Remove C++ SSL Certificate APIs
