+++ cpp/src/Ice/SSL/SSLEngine.cpp
@@ -81,86 +81,6 @@ Ice::SSL::SSLEngine::initialize()
 }
 
 void
-Ice::SSL::SSLEngine::verifyPeerCertName(const ConnectionInfoPtr& info, const string& address) const
-{
-    // For an outgoing connection, we compare the proxy address (if any) against fields in the server's certificate
-    // (if any).
-    if (_checkCertName && !info->certs.empty() && !address.empty())
-    {
-        const CertificatePtr cert = info->certs[0];
-
-        // Extract the IP addresses and the DNS names from the subject alternative names.
-        vector<pair<int, string>> subjectAltNames = cert->getSubjectAlternativeNames();
-        vector<string> ipAddresses;
-        vector<string> dnsNames;
-        for (vector<pair<int, string>>::const_iterator p = subjectAltNames.begin(); p != subjectAltNames.end(); ++p)
-        {
-            if (p->first == AltNAmeIP)
-            {
-                ipAddresses.push_back(IceUtilInternal::toLower(p->second));
-            }
-            else if (p->first == AltNameDNS)
-            {
-                dnsNames.push_back(IceUtilInternal::toLower(p->second));
-            }
-        }
-
-        bool certNameOK = false;
-        string addrLower = IceUtilInternal::toLower(address);
-        bool isIpAddress = IceInternal::isIpAddress(address);
-
-        // If address is an IP address, compare it to the subject alternative names IP address
-        if (isIpAddress)
-        {
-            certNameOK = find(ipAddresses.begin(), ipAddresses.end(), addrLower) != ipAddresses.end();
-        }
-        else
-        {
-            // If subjectAlt is empty compare it to the subject CN, otherwise compare it to the to the subject alt
-            // name dnsNames.
-            if (dnsNames.empty())
-            {
-                DistinguishedName d = cert->getSubjectDN();
-                string dn = IceUtilInternal::toLower(string(d));
-                string cn = "cn=" + addrLower;
-                string::size_type pos = dn.find(cn);
-                if (pos != string::npos)
-                {
-                    // Ensure we match the entire common name.
-                    certNameOK = (pos + cn.size() == dn.size()) || (dn[pos + cn.size()] == ',');
-                }
-            }
-            else
-            {
-                certNameOK = find(dnsNames.begin(), dnsNames.end(), addrLower) != dnsNames.end();
-            }
-        }
-
-        if (!certNameOK)
-        {
-            ostringstream ostr;
-            ostr << "SSL transport: ";
-            if (_verifyPeer > 0)
-            {
-                ostr << "ignoring ";
-            }
-            ostr << "certificate verification failure " << (isIpAddress ? "IP address mismatch" : "Hostname mismatch");
-            string msg = ostr.str();
-            if (_securityTraceLevel >= 1)
-            {
-                Trace out(getLogger(), _securityTraceCategory);
-                out << msg;
-            }
-
-            if (_verifyPeer > 0)
-            {
-                throw SecurityException(__FILE__, __LINE__, msg);
-            }
-        }
-    }
-}
-
-void
 Ice::SSL::SSLEngine::verifyPeer(const ConnectionInfoPtr& info) const
 {
     if (!_trustManager->verify(info))
