+++ cpp/src/Ice/SSL/SSLUtil.cpp
@@ -15,6 +15,7 @@
 #include "Ice/UniqueRef.h"
 #include "IceUtil/FileUtil.h"
 #include "IceUtil/StringUtil.h"
+#include "RFC2253.h"
 #include "SSLUtil.h"
 
 #include <fstream>
@@ -27,7 +28,7 @@ using namespace Ice::SSL;
 
 #if defined(__APPLE__)
 
-std::string
+string
 Ice::SSL::fromCFString(CFStringRef v)
 {
     string s;
@@ -153,3 +154,423 @@ Ice::SSL::checkPath(const string& path, const string& defaultDir, bool dir, stri
     }
     return false;
 }
+
+namespace
+{
+    const pair<string, string> certificateOIDS[] = {
+        {"2.5.4.3", "CN"},
+        {"2.5.4.4", "SN"},
+        {"2.5.4.5", "DeviceSerialNumber"},
+        {"2.5.4.6", "C"},
+        {"2.5.4.7", "L"},
+        {"2.5.4.8", "ST"},
+        {"2.5.4.9", "STREET"},
+        {"2.5.4.10", "O"},
+        {"2.5.4.11", "OU"},
+        {"2.5.4.12", "T"},
+        {"2.5.4.42", "G"},
+        {"2.5.4.43", "I"},
+        {"1.2.840.113549.1.9.8", "unstructuredAddress"},
+        {"1.2.840.113549.1.9.2", "unstructuredName"},
+        {"1.2.840.113549.1.9.1", "emailAddress"},
+        {"0.9.2342.19200300.100.1.25", "DC"}};
+    const int certificateOIDSSize = sizeof(certificateOIDS) / sizeof(pair<string, string>);
+}
+
+#ifdef ICE_USE_SECURE_TRANSPORT
+string
+Ice::SSL::certificateOIDAlias(const string& name)
+{
+    for (int i = 0; i < certificateOIDSSize; ++i)
+    {
+        const pair<string, string>* certificateOID = &certificateOIDS[i];
+        assert(certificateOID);
+        if (name == certificateOID->first)
+        {
+            return certificateOID->second;
+        }
+    }
+    return name;
+}
+#endif
+
+#if defined(ICE_USE_SCHANNEL)
+namespace
+{
+    string certNameToString(CERT_NAME_BLOB* certName)
+    {
+        assert(certName);
+        DWORD length = 0;
+        if ((length =
+                 CertNameToStr(X509_ASN_ENCODING, certName, CERT_OID_NAME_STR | CERT_NAME_STR_REVERSE_FLAG, 0, 0)) == 0)
+        {
+            throw CertificateEncodingException(__FILE__, __LINE__, IceUtilInternal::lastErrorToString());
+        }
+
+        vector<char> buffer(length);
+        if (!CertNameToStr(
+                X509_ASN_ENCODING,
+                certName,
+                CERT_OID_NAME_STR | CERT_NAME_STR_REVERSE_FLAG,
+                &buffer[0],
+                length))
+        {
+            throw CertificateEncodingException(__FILE__, __LINE__, IceUtilInternal::lastErrorToString());
+        }
+
+        string s(&buffer[0]);
+        for (int i = 0; i < certificateOIDSSize; ++i)
+        {
+            const pair<string, string>* certificateOID = &certificateOIDS[i];
+            assert(certificateOID);
+            const string name = string(certificateOID->first) + "=";
+            const string alias = string(certificateOID->second) + "=";
+            size_t pos = 0;
+            while ((pos = s.find(name, pos)) != string::npos)
+            {
+                s.replace(pos, name.size(), alias);
+            }
+        }
+        return s;
+    }
+}
+
+DistinguishedName
+Ice::SSL::getSubjectName(PCCERT_CONTEXT cert)
+{
+    return DistinguishedName(certNameToString(&cert->pCertInfo->Subject));
+}
+
+vector<pair<int, string>>
+Ice::SSL::getSubjectAltNames(PCCERT_CONTEXT cert)
+{
+    vector<pair<int, string>> altNames;
+
+    PCERT_EXTENSION extension =
+        CertFindExtension(szOID_SUBJECT_ALT_NAME2, cert->pCertInfo->cExtension, cert->pCertInfo->rgExtension);
+    if (extension)
+    {
+        CERT_ALT_NAME_INFO* altName;
+        DWORD length = 0;
+        if (!CryptDecodeObjectEx(
+                X509_ASN_ENCODING,
+                X509_ALTERNATE_NAME,
+                extension->Value.pbData,
+                extension->Value.cbData,
+                CRYPT_DECODE_ALLOC_FLAG,
+                0,
+                &altName,
+                &length))
+        {
+            throw CertificateEncodingException(__FILE__, __LINE__, IceUtilInternal::lastErrorToString());
+        }
+
+        for (DWORD i = 0; i < altName->cAltEntry; ++i)
+        {
+            CERT_ALT_NAME_ENTRY* entry = &altName->rgAltEntry[i];
+
+            switch (entry->dwAltNameChoice)
+            {
+                case CERT_ALT_NAME_RFC822_NAME:
+                {
+                    altNames.push_back(make_pair(AltNameEmail, Ice::wstringToString(entry->pwszRfc822Name)));
+                    break;
+                }
+                case CERT_ALT_NAME_DNS_NAME:
+                {
+                    altNames.push_back(make_pair(AltNameDNS, Ice::wstringToString(entry->pwszDNSName)));
+                    break;
+                }
+                case CERT_ALT_NAME_DIRECTORY_NAME:
+                {
+                    altNames.push_back(make_pair(AltNameDirectory, certNameToString(&entry->DirectoryName)));
+                    break;
+                }
+                case CERT_ALT_NAME_URL:
+                {
+                    altNames.push_back(make_pair(AltNameURL, Ice::wstringToString(entry->pwszURL)));
+                    break;
+                }
+                case CERT_ALT_NAME_IP_ADDRESS:
+                {
+                    if (entry->IPAddress.cbData == 4)
+                    {
+                        //
+                        // IPv4 address
+                        //
+                        ostringstream os;
+                        uint8_t* src = reinterpret_cast<uint8_t*>(entry->IPAddress.pbData);
+                        for (int j = 0; j < 4;)
+                        {
+                            int value = 0;
+                            uint8_t* dest = reinterpret_cast<uint8_t*>(&value);
+                            *dest = *src++;
+                            os << value;
+                            if (++j < 4)
+                            {
+                                os << ".";
+                            }
+                        }
+                        altNames.push_back(make_pair(AltNAmeIP, os.str()));
+                    }
+                    //
+                    // TODO IPv6 Address support.
+                    //
+                    break;
+                }
+                default:
+                {
+                    // Not supported
+                    break;
+                }
+            }
+        }
+        LocalFree(altName);
+    }
+    return altNames;
+}
+
+string
+Ice::SSL::encodeCertificate(PCCERT_CONTEXT cert)
+{
+    string s;
+    DWORD encodedLength = 0;
+    if (!CryptBinaryToString(
+            cert->pbCertEncoded,
+            cert->cbCertEncoded,
+            CRYPT_STRING_BASE64HEADER | CRYPT_STRING_NOCR,
+            0,
+            &encodedLength))
+    {
+        throw CertificateEncodingException(__FILE__, __LINE__, IceUtilInternal::lastErrorToString());
+    }
+
+    std::vector<char> encoded;
+    encoded.resize(encodedLength);
+    if (!CryptBinaryToString(
+            cert->pbCertEncoded,
+            cert->cbCertEncoded,
+            CRYPT_STRING_BASE64HEADER | CRYPT_STRING_NOCR,
+            &encoded[0],
+            &encodedLength))
+    {
+        throw CertificateEncodingException(__FILE__, __LINE__, IceUtilInternal::lastErrorToString());
+    }
+    s.assign(&encoded[0]);
+    return s;
+}
+#elif defined(ICE_USE_OPENSSL)
+
+//
+// Avoid old style cast warnings from OpenSSL macros
+//
+#    if defined(__GNUC__)
+#        pragma GCC diagnostic ignored "-Wold-style-cast"
+#        // Ignore OpenSSL 3.0 deprecation warning
+#        pragma GCC diagnostic ignored "-Wdeprecated-declarations"
+#    endif
+
+namespace
+{
+    string convertX509NameToString(X509_name_st* name)
+    {
+        BIO* out = BIO_new(BIO_s_mem());
+        X509_NAME_print_ex(out, name, 0, XN_FLAG_RFC2253);
+        BUF_MEM* p;
+        BIO_get_mem_ptr(out, &p);
+        string result = string(p->data, p->length);
+        BIO_free(out);
+        return result;
+    }
+
+    vector<pair<int, string>> convertGeneralNames(GENERAL_NAMES* gens)
+    {
+        vector<pair<int, string>> alt;
+        if (gens == 0)
+        {
+            return alt;
+        }
+        for (int i = 0; i < sk_GENERAL_NAME_num(gens); ++i)
+        {
+            GENERAL_NAME* gen = sk_GENERAL_NAME_value(gens, i);
+            pair<int, string> p;
+            p.first = gen->type;
+            switch (gen->type)
+            {
+                case GEN_EMAIL:
+                {
+                    ASN1_IA5STRING* str = gen->d.rfc822Name;
+                    if (str && str->type == V_ASN1_IA5STRING && str->data && str->length > 0)
+                    {
+                        p.second = string(reinterpret_cast<const char*>(str->data), str->length);
+                    }
+                    break;
+                }
+                case GEN_DNS:
+                {
+                    ASN1_IA5STRING* str = gen->d.dNSName;
+                    if (str && str->type == V_ASN1_IA5STRING && str->data && str->length > 0)
+                    {
+                        p.second = string(reinterpret_cast<const char*>(str->data), str->length);
+                    }
+                    break;
+                }
+                case GEN_DIRNAME:
+                {
+                    p.second = convertX509NameToString(gen->d.directoryName);
+                    break;
+                }
+                case GEN_URI:
+                {
+                    ASN1_IA5STRING* str = gen->d.uniformResourceIdentifier;
+                    if (str && str->type == V_ASN1_IA5STRING && str->data && str->length > 0)
+                    {
+                        p.second = string(reinterpret_cast<const char*>(str->data), str->length);
+                    }
+                    break;
+                }
+                case GEN_IPADD:
+                {
+                    ASN1_OCTET_STRING* addr = gen->d.iPAddress;
+                    // TODO: Support IPv6 someday.
+                    if (addr && addr->type == V_ASN1_OCTET_STRING && addr->data && addr->length == 4)
+                    {
+                        ostringstream ostr;
+                        for (int j = 0; j < 4; ++j)
+                        {
+                            if (j > 0)
+                            {
+                                ostr << '.';
+                            }
+                            ostr << static_cast<int>(addr->data[j]);
+                        }
+                        p.second = ostr.str();
+                    }
+                    break;
+                }
+                case GEN_OTHERNAME:
+                case GEN_EDIPARTY:
+                case GEN_X400:
+                case GEN_RID:
+                {
+                    //
+                    // TODO: These types are not supported. If the user wants
+                    // them, they have to get at the certificate data. Another
+                    // alternative is to DER encode the data (as the Java
+                    // certificate does).
+                    //
+                    break;
+                }
+            }
+            if (!p.second.empty())
+            {
+                alt.push_back(p);
+            }
+        }
+        sk_GENERAL_NAME_pop_free(gens, GENERAL_NAME_free);
+        return alt;
+    }
+}
+
+string
+Ice::SSL::getErrors(bool verbose)
+{
+    ostringstream ostr;
+
+    const char* file;
+    const char* data;
+    int line;
+    int flags;
+    unsigned long err;
+    int count = 0;
+    while ((err = ERR_get_error_line_data(&file, &line, &data, &flags)) != 0)
+    {
+        if (count > 0)
+        {
+            ostr << endl;
+        }
+
+        if (verbose)
+        {
+            if (count > 0)
+            {
+                ostr << endl;
+            }
+
+            char buf[200];
+            ERR_error_string_n(err, buf, sizeof(buf));
+
+            ostr << "error # = " << err << endl;
+            ostr << "message = " << buf << endl;
+            ostr << "location = " << file << ", " << line;
+            if (flags & ERR_TXT_STRING)
+            {
+                ostr << endl;
+                ostr << "data = " << data;
+            }
+        }
+        else
+        {
+            const char* reason = ERR_reason_error_string(err);
+            ostr << (reason == nullptr ? "unknown reason" : reason);
+            if (flags & ERR_TXT_STRING)
+            {
+                ostr << ": " << data;
+            }
+        }
+
+        ++count;
+    }
+
+    ERR_clear_error();
+
+    return ostr.str();
+}
+
+Ice::SSL::DistinguishedName
+Ice::SSL::getSubjectName(X509* certificate)
+{
+    return DistinguishedName(RFC2253::parseStrict(convertX509NameToString(X509_get_subject_name(certificate))));
+}
+
+std::vector<std::pair<int, string>>
+Ice::SSL::getSubjectAltNames(X509* certificate)
+{
+    return convertGeneralNames(
+        reinterpret_cast<GENERAL_NAMES*>(X509_get_ext_d2i(certificate, NID_subject_alt_name, 0, 0)));
+}
+
+string
+Ice::SSL::encodeCertificate(X509* certificate)
+{
+    BIO* out = BIO_new(BIO_s_mem());
+    int i = PEM_write_bio_X509(out, certificate);
+    if (i <= 0)
+    {
+        BIO_free(out);
+        throw CertificateEncodingException(__FILE__, __LINE__, getErrors(false));
+    }
+    BUF_MEM* p;
+    BIO_get_mem_ptr(out, &p);
+    string result = string(p->data, p->length);
+    BIO_free(out);
+    return result;
+}
+ICE_API X509*
+Ice::SSL::decodeCertificate(const string& data)
+{
+    BIO* cert = BIO_new_mem_buf(static_cast<void*>(const_cast<char*>(&data[0])), static_cast<int>(data.size()));
+    x509_st* x = PEM_read_bio_X509(cert, nullptr, nullptr, nullptr);
+    BIO_free(cert);
+    if (x == nullptr)
+    {
+        throw CertificateEncodingException(__FILE__, __LINE__, getErrors(false));
+    }
+    // Calling it with -1 for the side effects, this ensure that the extensions info is loaded
+    if (X509_check_purpose(x, -1, -1) == -1)
+    {
+        throw CertificateReadException(__FILE__, __LINE__, "error loading certificate:\n" + getErrors(false));
+    }
+    return x;
+}
+#endif
