+++ cpp/test/Glacier2/ssl/Server.cpp
@@ -2,6 +2,7 @@
 // Copyright (c) ZeroC, Inc. All rights reserved.
 //
 
+#include "../../src/Ice/SSL/SSLUtil.h"
 #include "Glacier2/PermissionsVerifier.h"
 #include "Glacier2/Session.h"
 #include "Ice/Ice.h"
@@ -52,17 +53,11 @@ public:
     {
         testContext(true, current.adapter->getCommunicator(), current.ctx);
 
-        auto cert = Ice::SSL::Certificate::decode(info.certs[0]);
+        auto cert = Ice::SSL::decodeCertificate(info.certs[0]);
         test(
-            cert->getIssuerDN() ==
-            Ice::SSL::DistinguishedName(
-                "emailAddress=info@zeroc.com,C=US,ST=Florida,L=Jupiter,O=ZeroC\\, Inc.,OU=Ice,CN=Ice Tests CA"));
-        test(
-            cert->getSubjectDN() ==
+            Ice::SSL::getSubjectName(cert) ==
             Ice::SSL::DistinguishedName(
                 "emailAddress=info@zeroc.com,C=US,ST=Florida,L=Jupiter,O=ZeroC\\, Inc.,OU=Ice,CN=client"));
-        test(cert->checkValidity());
-
         return true;
     }
 };
@@ -120,16 +115,11 @@ public:
 
         try
         {
-            auto cert = Ice::SSL::Certificate::decode(info.certs[0]);
+            auto cert = Ice::SSL::decodeCertificate(info.certs[0]);
             test(
-                cert->getIssuerDN() ==
-                Ice::SSL::DistinguishedName(
-                    "emailAddress=info@zeroc.com,C=US,ST=Florida,L=Jupiter,O=ZeroC\\, Inc.,OU=Ice,CN=Ice Tests CA"));
-            test(
-                cert->getSubjectDN() ==
+                Ice::SSL::getSubjectName(cert) ==
                 Ice::SSL::DistinguishedName(
                     "emailAddress=info@zeroc.com,C=US,ST=Florida,L=Jupiter,O=ZeroC\\, Inc.,OU=Ice,CN=client"));
-            test(cert->checkValidity());
         }
         catch (const Ice::SSL::CertificateReadException&)
         {
